refactor(bridge): extract the WorkspaceEdit host transform for reuse
Move the virtual-to-host WorkspaceEdit transformation core out of
rename.rs into bridge/protocol/workspace_edit.rs so the upcoming
textDocument/codeAction and workspace/applyEdit bridges (#568) can
reuse it. Rename keeps the JSON-RPC unwrapping wrapper and delegates
the coordinate work. Pure move, no behavior change.
