+++ src/lsp/bridge/text_document/rename.rs
@@ -10,22 +10,19 @@
 
 use std::io;
 
-use std::collections::HashMap;
-
 use crate::config::settings::BridgeServerConfig;
 use tower_lsp_server::ls_types::{
-    DocumentChangeOperation, DocumentChanges, NumberOrString, OneOf, Position, TextDocumentEdit,
-    TextEdit, Uri, WorkDoneProgressParams, WorkspaceEdit,
+    NumberOrString, Position, Uri, WorkDoneProgressParams, WorkspaceEdit,
 };
 use url::Url;
 
 use super::super::pool::{LanguageServerPool, UpstreamId};
 use tower_lsp_server::ls_types::RenameParams;
 
-use super::super::protocol::translate_virtual_range_to_host;
 use super::super::protocol::{
     JsonRpcRequest, RegionOffset, RequestId, VirtualDocumentUri,
     build_text_document_position_params, response_has_jsonrpc_error,
+    transform_workspace_edit_to_host,
 };
 
 impl LanguageServerPool {
@@ -115,10 +112,8 @@ fn build_rename_request(
 
 /// Transform a WorkspaceEdit response from virtual to host document coordinates.
 ///
-/// Per LSP spec a WorkspaceEdit may carry edits via `changes` (URI→TextEdit map) or
-/// `documentChanges`; both are handled. For each URI: real files pass through, the
-/// request's own virtual URI is translated, and other (cross-region) virtual URIs are
-/// filtered out.
+/// Unwraps the JSON-RPC response, then delegates the coordinate transformation
+/// to [`transform_workspace_edit_to_host`].
 fn transform_workspace_edit_response_to_host(
     mut response: serde_json::Value,
     request_virtual_uri: &str,
@@ -137,118 +132,9 @@ fn transform_workspace_edit_response_to_host(
     // Parse into typed WorkspaceEdit
     let mut edit: WorkspaceEdit = serde_json::from_value(result).ok()?;
 
-    // Transform changes map: { [uri: string]: TextEdit[] }
-    if let Some(changes) = &mut edit.changes {
-        transform_changes_map(changes, request_virtual_uri, host_uri, offset);
-    }
-
-    // Transform documentChanges array
-    if let Some(doc_changes) = &mut edit.document_changes {
-        transform_document_changes(doc_changes, request_virtual_uri, host_uri, offset);
-    }
+    transform_workspace_edit_to_host(&mut edit, request_virtual_uri, host_uri, offset);
 
     Some(edit)
-}
-
-/// Transform the `changes` map in a WorkspaceEdit.
-///
-/// Re-keys virtual URIs to host URI and transforms TextEdit ranges.
-/// Cross-region virtual URIs are removed entirely.
-fn transform_changes_map(
-    changes: &mut HashMap<Uri, Vec<TextEdit>>,
-    request_virtual_uri: &str,
-    host_uri: &Uri,
-    offset: &RegionOffset,
-) {
-    // Collect keys to process (can't modify HashMap keys in-place)
-    let keys: Vec<Uri> = changes.keys().cloned().collect();
-
-    for key in keys {
-        let uri_str = key.as_str();
-
-        // Case 1: Real file URI → keep as-is
-        if !VirtualDocumentUri::is_virtual_uri(uri_str) {
-            continue;
-        }
-
-        // Case 2: Same virtual URI → transform ranges, re-key to host URI
-        if uri_str == request_virtual_uri {
-            if let Some(mut edits) = changes.remove(&key) {
-                for edit in &mut edits {
-                    translate_virtual_range_to_host(&mut edit.range, offset);
-                }
-                changes.entry(host_uri.clone()).or_default().extend(edits);
-            }
-            continue;
-        }
-
-        // Case 3: Different virtual URI (cross-region) → filter out
-        changes.remove(&key);
-    }
-}
-
-/// Transform the `documentChanges` array in a WorkspaceEdit.
-///
-/// Handles both `Edits(Vec<TextDocumentEdit>)` and
-/// `Operations(Vec<DocumentChangeOperation>)` variants.
-/// File operations (CreateFile, RenameFile, DeleteFile) are preserved as-is.
-fn transform_document_changes(
-    doc_changes: &mut DocumentChanges,
-    request_virtual_uri: &str,
-    host_uri: &Uri,
-    offset: &RegionOffset,
-) {
-    match doc_changes {
-        DocumentChanges::Edits(edits) => {
-            edits.retain_mut(|edit| {
-                transform_text_document_edit(edit, request_virtual_uri, host_uri, offset)
-            });
-        }
-        DocumentChanges::Operations(ops) => {
-            ops.retain_mut(|op| match op {
-                DocumentChangeOperation::Edit(edit) => {
-                    transform_text_document_edit(edit, request_virtual_uri, host_uri, offset)
-                }
-                DocumentChangeOperation::Op(_) => true, // File operations preserved
-            });
-        }
-    }
-}
-
-/// Transform a single TextDocumentEdit's URI and edit ranges.
-///
-/// Returns `true` if the edit should be kept, `false` if it should be filtered out.
-fn transform_text_document_edit(
-    edit: &mut TextDocumentEdit,
-    request_virtual_uri: &str,
-    host_uri: &Uri,
-    offset: &RegionOffset,
-) -> bool {
-    let uri_str = edit.text_document.uri.as_str();
-
-    // Case 1: Real file URI → keep as-is
-    if !VirtualDocumentUri::is_virtual_uri(uri_str) {
-        return true;
-    }
-
-    // Case 2: Same virtual URI → transform
-    if uri_str == request_virtual_uri {
-        edit.text_document.uri = host_uri.clone();
-        // The version counted the virtual document; against the host URI it
-        // would make clients reject the edit as stale.
-        edit.text_document.version = None;
-        for one_of in &mut edit.edits {
-            let text_edit = match one_of {
-                OneOf::Left(text_edit) => text_edit,
-                OneOf::Right(annotated_edit) => &mut annotated_edit.text_edit,
-            };
-            translate_virtual_range_to_host(&mut text_edit.range, offset);
-        }
-        return true;
-    }
-
-    // Case 3: Cross-region → filter out
-    false
 }
 
 #[cfg(test)]
@@ -257,6 +143,7 @@ mod tests {
     use super::*;
     use rstest::rstest;
     use serde_json::json;
+    use tower_lsp_server::ls_types::{DocumentChanges, OneOf};
 
     // ==========================================================================
     // Rename request builder tests
